requirements.txt
fastapi
uvicorn
requests
PyMuPDF
python-docx
faiss-cpu
google-generativeai 